+++ CHANGES.md
@@ -5,6 +5,7 @@
 - The "Remove Events" context menu item in the Full Day view is only visible when events are available.
 - Added support for "Not (opposite)" searches in the "Search" dialog (with custom setting support included).
 - Added public function clearClipboard(), which clears the internal clipboard.
+- Added public function getClipboardEvent(), which returns the event details that have been copied to the internal clipboard.
 - 
 
 ### Version 1.3.6: